Extract test utils methods to own class.
