+++ libraries/trie/src/test/java/com/serwylo/lexica/trie/tests/CustomUsUkTrieTest.java
@@ -65,33 +65,6 @@ public class CustomUsUkTrieTest {
 		assertEverythingAboutTrie(trie);
 	}
 
-	@Test
-	public void testLetterMask() {
-		int actualMaskAbc = calcLetterMask("abc");
-		int actualMaskCde = calcLetterMask("CdE");
-		int actualMaskAllLower = calcLetterMask("abcdefghijklmnopqrstuvwxyz");
-		int actualMaskAllUpper = calcLetterMask("ABCDEFGHIJKLMNOPQRSTUVWXYZ");
-		int actualMaskAllMixed = calcLetterMask("aBcDeFgHiJkLmNoPqRsTuVwXyZ");
-
-		// TODO: Does "Qu" need any special treatment here? I suspect not.
-
-		int expectedMaskAbc = (1 << 0) | (1 << 1) | (1 << 2);
-		int expectedMaskCde = (1 << 2) | (1 << 3) | (1 << 4);
-		int expectedMaskAll =
-				(1 << 0) | (1 << 1) | (1 << 2) | (1 << 3) | (1 << 4) |
-				(1 << 5) | (1 << 6) | (1 << 7) | (1 << 8) | (1 << 9) |
-				(1 << 10) | (1 << 11) | (1 << 12) | (1 << 13) | (1 << 14) |
-				(1 << 15) | (1 << 16) | (1 << 17) | (1 << 18) | (1 << 19) |
-				(1 << 20) | (1 << 21) | (1 << 22) | (1 << 23) | (1 << 24) |
-				(1 << 25);
-
-		Assert.assertEquals(expectedMaskAbc, actualMaskAbc);
-		Assert.assertEquals(expectedMaskCde, actualMaskCde);
-		Assert.assertEquals(expectedMaskAll, actualMaskAllLower);
-		Assert.assertEquals(expectedMaskAll, actualMaskAllUpper);
-		Assert.assertEquals(expectedMaskAll, actualMaskAllMixed);
-	}
-
 	// "aeinqt" => "a" (all), "quit" (all), "aqua" (all), "queen" (uk)
 	// "abcehilmnor" => "america" (us), "monarch" (uk), "a" (all), "alibi" (all)
 	private static void assertEverythingAboutTrie(Trie trie) {
@@ -182,36 +155,14 @@ public class CustomUsUkTrieTest {
 	private static CompressedTrie deserialize(byte[] bytes, String availableLetters) {
 		try {
 			ByteArrayInputStream inputStream = new ByteArrayInputStream(bytes);
-			int letterMask = calcLetterMask(availableLetters);
-			int[] neighborMasks = calcNeighbourMasks();
+			int letterMask = Utils.calcLetterMask(availableLetters);
+			int[] neighborMasks = Utils.calcNeighbourMasks();
 
 			return new CompressedTrie(inputStream, letterMask, neighborMasks, true, true);
 		} catch (IOException e) {
 			Assert.fail();
 			return null;
 		}
-	}
-
-	/**
-	 * The way {@link CompressedTrie}s are stored in memory after being read from disk is to
-	 * only include words that can be made from the letters available.
-	 */
-	private static int calcLetterMask(String availableLetters) {
-		int mask = 0;
-		for (int i = 0; i < availableLetters.length(); i ++) {
-			char character = availableLetters.charAt(i);
-			mask |= 1 << Trie.ctoi(character);
-		}
-		return mask;
-	}
-
-	// TODO: Be more specific about valid transitions in future tests.
-	private static int[] calcNeighbourMasks() {
-		int[] neighbourMasks = new int[26];
-		for (int i = 0; i < 26; i ++) {
-			neighbourMasks[i] = 0xffffffff;
-		}
-		return neighbourMasks;
 	}
 
 }
